backend/api/requirements.txt
mysql-connector-python==9.1.0
Flask==3.1.0
Flask-RESTful==0.3.10
requests==2.32.3